+++ src/components/main_page/Hero.vue
